+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 ----------------
 
 * Fixed loading .ipuz puzzles that do not have a solution element.
+* Fixed loading .ipuz puzzles that do not have Across and/or Down clues.
 
 0.1 (August 17, 2014)
 ---------------------